+++ src/pouch.utils.js
@@ -322,136 +322,6 @@ var rootToLeaf = function(tree) {
   return paths;
 };
 
-var ajax = function ajax(options, callback) {
-  if (typeof options === "function") {
-    callback = options;
-    options = {};
-  }
-  var defaultOptions = {
-    method : "GET",
-    headers: {},
-    json: true,
-    processData: true,
-    timeout: 10000
-  };
-  options = extend(true, defaultOptions, options);
-  if (options.auth) {
-      var token = btoa(options.auth.username + ':' + options.auth.password);
-      options.headers.Authorization = 'Basic ' + token;
-  }
-  var onSuccess = function(obj, resp, cb){
-    if (!options.binary && !options.json && options.processData && typeof obj !== 'string') {
-      obj = JSON.stringify(obj);
-    } else if (!options.binary && options.json && typeof obj === 'string') {
-      try {
-        obj = JSON.parse(obj);
-      } catch (e) {
-        // Probably a malformed JSON from server
-        call(cb, e);
-        return;
-      }
-    }
-    call(cb, null, obj, resp);
-  };
-  var onError = function(err, cb){
-    var errParsed;
-    var errObj = err.responseText ? {status: err.status} : err; //this seems too clever
-    try{
-      errParsed = JSON.parse(err.responseText); //would prefer not to have a try/catch clause
-      errObj = extend(true, {}, errObj, errParsed);
-    } catch(e){}
-    call(cb, errObj);
-  };
-  if (typeof window !== 'undefined' && window.XMLHttpRequest) {
-    var timer,timedout  = false;
-    var xhr = new XMLHttpRequest();
-    xhr.open(options.method, options.url);
-    if (options.json) {
-      options.headers.Accept = 'application/json';
-      options.headers['Content-Type'] = options.headers['Content-Type'] || 'application/json';
-      if (options.body && options.processData && typeof options.body !== "string") {
-        options.body = JSON.stringify(options.body);
-      }
-    }
-    if (options.binary) {
-      xhr.responseType = 'arraybuffer';
-    }
-    for (var key in options.headers){
-      xhr.setRequestHeader(key, options.headers[key]);
-    }
-    if (!("body" in options)) {
-      options.body = null;
-    }
-
-    var abortReq = function() {
-        timedout=true;
-        xhr.abort();
-        call(onError, xhr, callback);
-    };
-    xhr.onreadystatechange = function() {
-      if (xhr.readyState !== 4 || timedout) {
-        return;
-      }
-      clearTimeout(timer);
-      if (xhr.status >= 200 && xhr.status < 300) {
-        var data;
-        if (options.binary) {
-          data = new Blob([xhr.response || ''], {type: xhr.getResponseHeader('Content-Type')});
-        } else {
-          data = xhr.responseText;
-        }
-        call(onSuccess, data, xhr, callback);
-      } else {
-         call(onError, xhr, callback);
-      }
-    };
-    if (options.timeout > 0) {
-      timer = setTimeout(abortReq, options.timeout);
-    }
-    xhr.send(options.body);
-    return {abort:abortReq};
-  } else {
-    if (options.json) {
-      if (!options.binary) {
-        options.headers.Accept = 'application/json';
-      }
-      options.headers['Content-Type'] = options.headers['Content-Type'] || 'application/json';
-    }
-    if (options.binary) {
-      options.encoding = null;
-      options.json = false;
-    }
-    if (!options.processData) {
-      options.json = false;
-    }
-    return request(options, function(err, response, body) {
-      if (err) {
-        err.status = response ? response.statusCode : 400;
-        return call(onError, err, callback);
-      }
-
-      var content_type = response.headers['content-type'];
-      var data = (body || '');
-
-      // CouchDB doesn't always return the right content-type for JSON data, so
-      // we check for ^{ and }$ (ignoring leading/trailing whitespace)
-      if (!options.binary && (options.json || !options.processData) && typeof data !== 'object' &&
-          (/json/.test(content_type) ||
-           (/^[\s]*\{/.test(data) && /\}[\s]*$/.test(data)))) {
-        data = JSON.parse(data);
-      }
-      
-      if (data.error) {
-        data.status = response.statusCode;
-        call(onError, data, callback);
-      }
-      else {
-        call(onSuccess, data, response, callback);
-      }
-    });
-  }
-};
-
 // Basic wrapper for localStorage
 var win = this;
 var localJSON = (function(){
@@ -487,6 +357,8 @@ if (typeof module !== 'undefined' && module.exports) {
     }
   };
   var extend = require('./deps/extend');
+  var ajax = require('./deps/ajax');
+
   request = require('request');
   _ = require('underscore');
   $ = _;
@@ -508,7 +380,6 @@ if (typeof module !== 'undefined' && module.exports) {
     writeCheckpoint: writeCheckpoint,
     arrayFirst: arrayFirst,
     filterChange: filterChange,
-    ajax: ajax,
     atob: function(str) {
       return decodeURIComponent(escape(new Buffer(str, 'base64').toString('binary')));
     },
@@ -516,6 +387,7 @@ if (typeof module !== 'undefined' && module.exports) {
       return new Buffer(unescape(encodeURIComponent(str)), 'binary').toString('base64');
     },
     extend: extend,
+    ajax: ajax,
     traverseRevTree: traverseRevTree,
     rootToLeaf: rootToLeaf,
   };
